+++ NEWS.md
@@ -69,6 +69,7 @@ Standard library changes
 * On aarch64, OpenBLAS now uses an ILP64 BLAS like all other 64-bit platforms. ([#39436])
 * OpenBLAS is updated to 0.3.13. ([#39216])
 * SuiteSparse is updated to 5.8.1. ([#39455])
+* `cis(A)` now supports matrix arguments ([#40194]).
 
 #### Markdown